requirements.txt
et-xmlfile==1.1.0
mysql-connector-python==8.3.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
prettytable==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tqdm==4.66.2
tzdata==2024.1
wcwidth==0.2.13
